Modify the machine_nodeb table structure

+++ src/app/server/service/machine/machinemanager.go
@@ -1,9 +1,21 @@
 package machine
 
 import (
+	"errors"
+
 	departservice "gitee.com/openeuler/PilotGo/app/server/service/depart"
+	"gitee.com/openeuler/PilotGo/global"
+	"gitee.com/openeuler/PilotGo/utils"
 
 	"gitee.com/openeuler/PilotGo/app/server/service/internal/dao"
+)
+
+// 机器状态
+const (
+	OnlineStatus      = "online"
+	OfflineStatus     = "offline"
+	NormalStauts      = "normal"
+	MaintenanceStatus = "maintenance"
 )
 
 type MachineNode = dao.MachineNode
@@ -16,11 +28,15 @@ type DeleteUUID struct {
 	Deluuid []string `json:"deluuid"`
 }
 
-func MachineInfo(depart *Depart, offset, size int) (int64, []dao.Res, error) {
+// 新增agent机器
+func AddMachine(Machine *MachineNode) error {
+	return Machine.Add()
+}
 
+func MachineInfo(depart *Depart, offset, size int) (int64, []dao.Res, error) {
 	var TheDeptAndSubDeptIds []int
 	departservice.ReturnSpecifiedDepart(depart.ID, &TheDeptAndSubDeptIds)
-	TheDeptAndSubDeptIds = append(TheDeptAndSubDeptIds, depart.ID)
+	TheDeptAndSubDeptIds = append([]int{depart.ID}, TheDeptAndSubDeptIds...)
 	total, data, err := dao.GetMachinePaged(TheDeptAndSubDeptIds, offset, size)
 	return total, data, err
 }
@@ -30,7 +46,7 @@ func ReturnMachinePaged(departid, offset, size int) (int64, []dao.Res, error) {
 }
 
 func MachineAllData() ([]map[string]string, error) {
-	AllData, err := dao.MachineAllData()
+	AllData, err := dao.MachineAll()
 	if err != nil {
 		return nil, err
 	}
@@ -41,25 +57,32 @@ func MachineAllData() ([]map[string]string, error) {
 	return datas, nil
 }
 
-func Machines() ([]dao.Res, error) {
-	AllData, err := dao.MachineAllData()
-	if err != nil {
-		return nil, err
-	}
-	return AllData, nil
+func MachineAll() ([]dao.Res, error) {
+	AllData, err := dao.MachineAll()
+	return AllData, err
 }
 
+// 删除机器，删除之前先校验uuid是否存在
 func DeleteMachine(Deluuid []string) map[string]string {
 	machinelist := make(map[string]string)
 	for _, machinedeluuid := range Deluuid {
-		if err := dao.DeleteMachine(machinedeluuid); err != nil {
+		node, err := MachineInfoByUUID(machinedeluuid)
+		if err != nil {
 			machinelist[machinedeluuid] = err.Error()
+		}
+		if node.ID != 0 {
+			if err := dao.DeleteMachine(machinedeluuid); err != nil {
+				machinelist[machinedeluuid] = err.Error()
+			}
+		} else {
+			machinelist[machinedeluuid] = errors.New("该机器不存在").Error()
 		}
 	}
 	return machinelist
 }
 
-func MachineBasic(uuid string) (*Res, error) {
+// 根据uuid查找机器信息
+func MachineInfoByUUID(uuid string) (*Res, error) {
 	node, err := dao.MachineInfoByUUID(uuid)
 	if err != nil {
 		return nil, err
@@ -67,37 +90,48 @@ func MachineBasic(uuid string) (*Res, error) {
 
 	depart, err := dao.GetDepartById(node.DepartId)
 	r := &Res{
-		ID:         node.ID,
-		Departid:   node.DepartId,
-		Departname: depart.Depart,
-		IP:         node.IP,
-		UUID:       node.MachineUUID,
-		CPU:        node.CPU,
-		State:      node.State,
-		Systeminfo: node.Systeminfo,
+		ID:          node.ID,
+		Departid:    node.DepartId,
+		Departname:  depart.Depart,
+		IP:          node.IP,
+		UUID:        node.MachineUUID,
+		CPU:         node.CPU,
+		Runstatus:   node.RunStatus,
+		Maintstatus: node.MaintStatus,
+		Systeminfo:  node.Systeminfo,
 	}
 	return r, err
 }
 
-func UpdateMachineState(uuid string, state int) error {
-	return dao.UpdateMachineState(uuid, state)
+func UpdateMachine(uuid string, ma *MachineNode) error {
+	return dao.UpdateMachine(uuid, ma)
 }
 
-func IsUUIDExist(uuid string) (bool, error) {
-	return dao.IsUUIDExist(uuid)
-}
-
-// 根据uuid获取部门id
-func UUIDForDepartId(uuid string) (int, error) {
-	return dao.UUIDForDepartId(uuid)
-}
-
-// 更新机器IP及状态
-func UpdateMachineIPState(uuid, ip string, state int) error {
-	return dao.UpdateMachineIPState(uuid, ip, state)
-}
-
-// 新增agent机器
-func AddNewMachine(Machine MachineNode) error {
-	return dao.AddNewMachine(Machine)
+func ModifyMachineDepart(MachineID string, DepartID int) error {
+	//查询部门节点是否存在
+	depart, err := dao.GetDepartById(DepartID)
+	if err != nil {
+		return err
+	}
+	if depart.ID == 0 {
+		return errors.New("此部门不存在")
+	}
+	ResIds := utils.String2Int(MachineID)
+	for _, id := range ResIds {
+		machine, err := dao.MachineInfo(id)
+		if err != nil {
+			return err
+		}
+		Ma := &dao.MachineNode{DepartId: DepartID}
+		if DepartID != global.UncateloguedDepartId {
+			Ma.MaintStatus = NormalStauts
+		} else {
+			Ma.MaintStatus = MaintenanceStatus
+		}
+		err = dao.UpdateMachine(machine.MachineUUID, Ma)
+		if err != nil {
+			return err
+		}
+	}
+	return nil
 }
